Create simple keylogger
Explanation:
Importing Modules: The pynput.keyboard module is imported to monitor keyboard events.
on_press Function: This function is called whenever a key is pressed. It tries to print and log the alphanumeric key or special key pressed.
on_release Function: This function is called whenever a key is released. It stops the listener if the Esc key is pressed.
Listener: The Listener context manager collects keyboard events and calls the respective functions (on_press and on_release).
Usage:
Run the script in the background.
All keystrokes will be logged to keylog.txt.
Press the Esc key to stop the keylogger.